Relationships contentful

See merge request bitdarlene/darlene-backend!30

+++ backend/prereqs/utils.py
@@ -3,29 +3,32 @@ from backend.models import Badge, ActivityBadgePrereqs, ModuleBadgePrereqs, Topi
 
 
 # Function that creates a badge_reqs depending on the object_type
-def assign_badge_prereqs(badge_data, selected_object, object_type):
-    for badge_info in badge_data:
-        xp = badge_info["xp"]
-        badge = Badge.query.get(badge_info["id"])
+def assign_badge_prereqs(contentful_data, selected_object, object_type):
+    if "badge_prereqs" in contentful_data["parameters"]:
+        badge_list = contentful_data["parameters"]["badge_prereqs"]["en-US"]["badge_prereqs"]
 
-        # If the badge exists, then add the badge to the object
-        if badge:
-            target_badge = None
+        for badge_info in badge_list:
+            xp = badge_info["xp"]
+            badge = Badge.query.get(badge_info["badge_id"])
 
-            if object_type == "Activity":
-                target_badge = ActivityBadgePrereqs(xp=xp)
-                target_badge.badge = badge
-                target_badge.activity_id = selected_object.id
-            elif object_type == "Module":
-                target_badge = ModuleBadgePrereqs(xp=xp)
-                target_badge.badge = badge
-                target_badge.module_id = selected_object.id
-            elif object_type == "Topic":
-                target_badge = TopicBadgePrereqs(xp=xp)
-                target_badge.badge = badge
-                target_badge.topic_id = selected_object.id
+            # If the badge exists, then add the badge to the object
+            if badge:
+                target_badge = None
 
-            selected_object.badge_prereqs.append(target_badge)
+                if object_type == "Activity":
+                    target_badge = ActivityBadgePrereqs(xp=xp)
+                    target_badge.badge = badge
+                    target_badge.activity_id = selected_object.id
+                elif object_type == "Module":
+                    target_badge = ModuleBadgePrereqs(xp=xp)
+                    target_badge.badge = badge
+                    target_badge.module_id = selected_object.id
+                elif object_type == "Topic":
+                    target_badge = TopicBadgePrereqs(xp=xp)
+                    target_badge.badge = badge
+                    target_badge.topic_id = selected_object.id
+
+                selected_object.badge_prereqs.append(target_badge)
 
     return
 
@@ -33,8 +36,9 @@ def assign_badge_prereqs(badge_data, selected_object, object_type):
 # Function to delete badge prereqs based on the selected object
 # selected object could be an Activity, Module
 def delete_badge_prereqs(selected_object):
-    for badge in selected_object.badge_prereqs:
-        db.session.delete(badge)
-    db.session.commit()
+    if selected_object.badge_prereqs:
+        for badge in selected_object.badge_prereqs:
+            db.session.delete(badge)
+        db.session.commit()
 
     return
